Modification to _converted.csv file header records
Based on feedback from Joseph and a closer inspection of 03401.cnv.csv in the repo the default behavior for how converted data is written to csv files was changed from a 1-line header to a 2-line header.

The first line of the header is the data type (pressure, temp, etc) '_' and the units of measure (mBar, C, etc)
The second line of hte heard is the dtype to use when importing the data followed by '_' (int_, float_, bool_, datetime_)

The converted data import script (run.py) has been updated to accept this new format.

A new function (saveConvertedDataToFile()) has been added to converter_scaffolding.py.
This function requires a dataframe, filename and debug flag as arguments.  The function will write the dataframe to csv file with the new 2-line header format.

An updated version of the bottle.py (bottleNewStyle.py) has been added.  This file replicates the functions of bottle.py but utilizes the cnv import/export architecture


Former-commit-id: bf7a5c1e8b73debd59f5b772afad43be6da455d0
Former-commit-id: d44ef48e436c6b95b56fd499ded0be9b5981c0f2